style: add style to `Home`
